Refactored AATA code into extracters.aata_data.
