+++ aata.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+# TODO: refactor code from knoedler files that is actually just linkedart related
 # TODO: ensure that multiple serializations to the same uuid are merged. e.g. a journal article with two authors, that each get asserted as carrying out the creation event.
 # TODO: move xml-related code to just the first extract node. all other nodes should be based on a resulting dict to allow re-use with non-xml source data.
 
@@ -11,13 +12,13 @@ from bonobo.nodes import Limit
 from bonobo.config import Configurable, Option
 import itertools
 import bonobo_sqlalchemy
-import lxml.etree
 import sqlite3
 
 from cromulent import model, vocab
-from extracters.xml import XMLReader, ExtractXPath, FilterXPathEqual, print_xml_element_text
+from extracters.xml import XMLReader
 from extracters.basic import AddArchesModel, AddFieldNames, Serializer, deep_copy, Offset, add_uuid, Trace
-from extracters.knoedler_data import *
+from extracters.aata_data import make_aata_article_dict, make_aata_authors, make_aata_abstract, add_aata_object_type
+from extracters.knoedler_data import add_person_aat_labels
 from extracters.knoedler_linkedart import *
 from extracters.arches import ArchesWriter, FileWriter
 from settings import *
@@ -49,125 +50,17 @@ else:
 	# WRITER	= ArchesWriter()
 
 
-def make_author_dict(e, parent_object, parent_data, event):
-	aid = e.find('./author_id')
-	if aid is not None:
-		name = aid.findtext('display_term')
-		auth_id = aid.findtext('gaia_auth_id')
-		auth_type = aid.findtext('gaia_auth_type')
-		yield {
-			'_source_element': e,
-			'parent': parent_object,
-			'parent_data': parent_data,
-			'label': name,
-			'events': [event],
-			'uid': 'AATA-P-%s-%s-%s' % (auth_type, auth_id, name)
-		}
-	else:
-		print('*** No author_id found for record %s:' % (parent,))
-		print(lxml.etree.tostring(e).decode('utf-8'))
-
-def make_article_dict(e):
-	doc_type = e.findtext('./record_desc_group/doc_type')
-	title_type = e.findtext('./title_group/title_type')
-	title = e.findtext('./title_group[title_type = "Analytic"]/title')
-	rid = e.findtext('./record_id_group/record_id')
-	return {
-		'_source_element': e,
-		'label': title,
-		'uid': 'AATA-%s-%s-%s' % (doc_type, rid, title)
-	}
-
 class AddDataDependentArchesModel(Configurable):
 	models = Option()
 	def __call__(self, data):
 		data['_ARCHES_MODEL'] = self.models['LinguisticObject']
 		return data
 
-def add_object_type(data):
-	doc_types = { # should this be in settings (or elsewhere)?
-		'AV': vocab.AudioVisualContent,
-		'BA': vocab.Chapter,
-		'BC': vocab.Monograph, # TODO: is this right?
-		'BM': vocab.Monograph,
-		'JA': vocab.Article,
-		'JW': vocab.Issue,
-		'PA': vocab.Patent,
-		'TH': vocab.Thesis,
-		'TR': vocab.TechnicalReport
-	}
-	e = data['_source_element']
-	atype = e.findtext('./record_desc_group/doc_type')
-	data['object_type'] = doc_types[atype]
-	return data
-
-def make_authors(data):
-	e = data['_source_element']
-	object = data['_LOD_OBJECT']
-	event = model.Creation()
-	object.created_by = event
-	for ag in e.xpath('./authorship_group'):
-		subevent = model.Creation()
-		event.part = subevent
-		role = ag.findtext('author_role')
-		if role is None:
-			print('*** No author role found for authorship group in %s:' % (object,))
-			print(lxml.etree.tostring(ag).decode('utf-8'))
-		subevent._label = role
-		for a in ag.xpath('./author'):
-			yield from make_author_dict(a, parent_object=object, parent_data=data, event=subevent)
-
 def make_la_record(data: dict):
 	otype = data['object_type']
 	object = otype(ident="urn:uuid:%s" % data['uuid'])
 	object._label = data['label']
 	return add_crom_data(data=data, what=object)
-
-def language_object_from_code(code):
-	languages = {
-		'eng': {'ident': 'http://vocab.getty.edu/aat/300388277', 'label': 'English'}
-	}
-	try:
-		kwargs = languages[code]
-		return model.Language(**kwargs)
-	except KeyError:
-		print('*** No AAT link for language %r' % (language,))
-
-def make_abstract(data):
-	author_data = data
-	data = author_data['parent_data']
-	e = data['_source_element']
-	object = data['_LOD_OBJECT']
-	for ag in e.xpath('./abstract_group'):
-		abstract = model.LinguisticObject()
-		a = ag.find('./abstract')
-		author_abstract_flag = ag.findtext('./author_abstract_flag')
-		if a is not None:
-			if author_abstract_flag == 'yes':
-				event = model.Creation()
-				abstract.created_by = event
-				event.carried_out_by = author_data['_LOD_OBJECT']
-			else:
-				# TODO: what about the creation event when author_abstract_flag != 'yes'?
-				pass
-
-			content = a.text
-			language = a.attrib.get('lang')
-			abstract.content = content
-			abstract.classified_as = model.Type(ident='http://vocab.getty.edu/aat/300026032', label='Abstract') # TODO: is this the right aat URI?
-			if language is not None:
-				l = language_object_from_code(language)
-				if l is not None:
-					abstract.language = l
-
-			abstract.refers_to = object
-			yield {
-				'_LOD_OBJECT': abstract,
-				'_source_element': ag,
-				'parent': object,
-				'content': content,
-				'uid': 'AATA-A-%s-%s' % (data['uid'], content)
-			}
 
 def make_la_abstract(data: dict):
 	return add_crom_data(data=data, what=data['_LOD_OBJECT'])
@@ -180,9 +73,9 @@ def get_graph(files, **kwargs):
 		articles = graph.add_chain(
 			aata_records,
 			Limit(LIMIT),
-			make_article_dict,
+			make_aata_article_dict,
 			add_uuid,
-			add_object_type,
+			add_aata_object_type,
 			make_la_record,
 			AddDataDependentArchesModel(models=arches_models),
 		)
@@ -196,7 +89,7 @@ def get_graph(files, **kwargs):
 			)
 		
 		people = graph.add_chain(
-			make_authors,
+			make_aata_authors,
 			AddArchesModel(model=arches_models['Person']),
 			add_uuid,
 			add_person_aat_labels,
@@ -213,7 +106,7 @@ def get_graph(files, **kwargs):
 			)
 
 		abstracts = graph.add_chain(
-			make_abstract,
+			make_aata_abstract,
 			AddArchesModel(model=arches_models['LinguisticObject']),
 			add_uuid,
 			make_la_abstract,
@@ -227,13 +120,6 @@ def get_graph(files, **kwargs):
 				WRITER,
 				_input=abstracts.output
 			)
-
-# 		graph.add_chain(
-# 			make_publishers,
-# 			SRLZ,
-# 			WRITER,
-# 			_input=articles.output
-# 		)
 
 	return graph
